requirements.txt
aiofiles==23.2.1
aiohttp==3.7.4.post0
py-cord==2.4.1
patool==1.12
Requests==2.31.0
